requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
alabaster=0.7.12=pypi_0
async-timeout=4.0.3=py39h06a4308_0
attrs=23.2.0=pyh71513ae_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
blinker=1.7.0=pyhd8ed1ab_0
boto3=1.34.14=pypi_0
botocore=1.34.14=pypi_0
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.2.2=hbcca054_0
cachetools=5.2.0=pypi_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
contourpy=1.0.5=pypi_0
cryptography=37.0.1=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
deeprobust=0.2.5=pypi_0
distlib=0.3.6=pypi_0
faiss-gpu=1.7.3=py3.9_h28a55e0_0_cuda11.3
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
filelock=3.8.0=pypi_0
fonttools=4.37.3=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.4.0=py39h5eee18b_0
gensim=3.8.3=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.14.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.50.0=pypi_0
h5py=3.7.0=pypi_0
huggingface-hub=0.9.1=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.22.0=pypi_0
importlib-metadata=5.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libfaiss=1.7.3=hfc2d529_0_cuda11.3
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
littleutils=0.2.2=pypi_0
llvmlite=0.39.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.6.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.4=py39h5eee18b_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
networkx=2.8.6=pypi_0
numba=0.56.2=pypi_0
numpy=1.22.3=py39he7a7128_0
numpy-base=1.22.3=py39hf524024_0
oauthlib=3.2.2=pyhd8ed1ab_0
ogb=1.3.4=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
outdated=0.2.1=pypi_0
packaging=21.3=pypi_0
pandas=1.5.0=pypi_0
pillow=9.2.0=py39hace64e9_1
pip=23.3.1=py39h06a4308_0
platformdirs=2.5.2=pypi_0
protobuf=3.20.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.1.0=py39_torch_1.12.0_cu113
pyjwt=2.8.0=pyhd8ed1ab_1
pymetis=2022.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_1
python-dateutil=2.8.2=pypi_0
python_abi=3.9=2_cp39
pytorch=1.12.0=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-cluster=1.6.0=py39_torch_1.12.0_cu113
pytorch-mutex=1.0=cuda
pytorch-pretrained-bert=0.6.2=pypi_0
pytorch-scatter=2.0.9=py39_torch_1.12.0_cu113
pytorch-sparse=0.6.15=py39_torch_1.12.0_cu113
pytz=2022.2.1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.9.13=pypi_0
requests=2.28.1=py39h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
s3transfer=0.10.0=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.2=py39h6a678d5_0
scipy=1.9.1=pypi_0
setuptools=59.8.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smart-open=6.2.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.38.5=hc218d9a_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorboardx=2.5.1=pypi_0
texttable=1.6.4=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.8.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
torchaudio=0.12.0=py39_cu113
torchvision=0.13.0=py39_cu113
tqdm=4.64.0=pypi_0
transformers=4.21.0=pypi_0
typer=0.6.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39h06a4308_0
virtualenv=20.16.2=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wikipedia=1.4.0=pypi_0
wrapt=1.14.1=pypi_0
xz=5.2.5=h7f8727e_1
yarl=1.7.2=py39hb9d737c_2
zipp=3.8.1=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
